Load chunks around player spawn point instead of world center
We have a chicken-and-egg problem with player spawning and chunk loading
because:

1: Chunk in which player need to be spawned in need to be loaded first
   before the player could be spawned.
2: The only way a chunk is loaded if if it is near a spawned player.

To solve problem, we previously unconditionally load chunks at a fixed
radius around the world center. This is restrictive since if the initial
spawn point of the player is not inside the chunks that are
unconditionally loaded near the world center, we would face the same
chicken-and-egg dilemma.

The solution is to unconditionally load chunks around player spawn point
instead of world center if we are unable to spawn the player.