Fix race condition in theaad pool
We no longer call job->destroy after job->invoke. Currently, the only
place we use the thread pool is in our chunk generator in which we have
wrapper struct that is simultaneously in a hash table and the job queue
of the thread pool. The lifetime of the wrapper struct is managed by the
hash table and not the job queue, and we indicate the fact that a chunk
has been generated by doing an atomic store to a pointer with release
memory ordering in job->invoke. After that, the main thread is free to
steal the chunk pointer and free the wrapper struct. Hence, it is a
race condition if we try to call wrapper->job->destroy even if we do
nothing in the callback since we are accessing a function pointer in the
wrapper struct, which may have been freed.