Hacky entity database implementation with sqlite.
We used to store each entity in a separate file. We are wasting a lot of
storage space by doing so, since file system usually allocates space in
blocks which can be as huge as 4KiB. However, the size of our entity is
nowhere near that. Some file systems have feature called block
suballocation which can help with that, but we should not be relying on
that, since that is not a common feature.

Another hope is that by utilizing storage spacce better, we would cut
down on unnecessary io and improve performance. Unfortunately, it does
not seem of have much of a effect in that regard.