libnet-server: Add missing removal of client from intererested list
During update, a client may be added to the interested list if not
already because we attempt to send data to it such as if a new chunk has
been generated. The problem is that we do not process the interested
list until we process all epoll events. Between update and the
processing of the interested list, the client socket may become
readable/writable and we will need to handle it. The problem is that the
connection may become closed, in which case we need to destroy the
client object in our server. The missing piece of code is that we also
need to unlink the client object from the interested list or else we
will have a dangling pointer.