IO uring implementation for chunk database
It turns out that posix aio kinda sucks because there is no aio_open(3).
Before we even try to call aio_read(3), we need a open the file with
open(2), which surprise, surprise is a blocking system call as it need
to traverse the filesystem structure to check if the file exist and we
have the right permission to access it.