Restructure the chunk system
Before, the chunk system manage both which chunks is considered "active"
and loading and saving of blocks inside a chunk, and the entity system
depends on the chunk system in deciding on which entity to load/save.
Pictorially,

  ----------------    -----------------
  | chunk system | <-- | entity system |
  ----------------    -----------------

We separate the management of blocks inside chunk out of the chunk
system and called the collection of blocks inside a single chunk "block
group". Pictorially,

  ----------------     -----------------
  | chunk system | <-- | entity system |
  ----------------  |  -----------------
                    |
                    |  ----------------
                    ---| block system |
                       ----------------

Is there any technical reasons to do so? Probably not but who cares?