Fix function prototype of thread_pool_func()
Add missing argument "void *data" to the function prototype of
thread_pool_func. This should not have been necessary (or at least it
have been working and is still working with clang) since in C - a
function with no argument should instead be treated as a function
without a prototype, and thus be able to accept any arguments. Instead,
a function without argument should be declared like "void
thread_pool_func(void)".