Also remesh a block group if its render info is missing
This can happen if a block group is remeshed but then its render info is
later discarded because it temporarily go outside of render distance. In
that case, the remesh flag on block group will not be reset.

This is kinda hacky in the sense that ideally, the remesh flag should be
attached to render infos instead of block groups.