Wayland-style message passing for libnet server implementation
Looking at profiling output in the past, there have always been gaps
between consecutive server updates which are the time used for network
socket handling. This is the case even if we are already taking over our
target delta time as specified FIXED_DT in the current frame, which is
not ideal. The solution is to put the responsiblity of network socket
handling on a separate thread and communicate with main thread using the
requests/events model and the idea of acknowledgement requests/evente to
synchronize state from wayland.

A stb_ds style ring buffer and proper time handling primitive that does
not depend on glfw has also been implemented in libcore.

An added benenfit is that we use poll(2) instead level-triggered
epoll(7) because we do not expect to have high number of concurrent
connections and this avoid running into starvation problem with our
previous implementation. (It is probably technically possible to avoid
such a problem with level-triggered epoll(7) but oh well I have not
implemented it before it is to be replaced now).