Introduce chunk regions and major/minor chunks
Instead of allowing mod to specifiy which chunks are active
individually, they now need to be specified in terms of spherical
regions, which are much easier to iterate with and it is also much
easier to check if a chunk is active. This cuts down on a lot of the
hash table accesses we have previously which are slow even though they
are supposedly amortized O(1). As it turns out, the hidden constant can
be huge but that may also be due to us switching to stbds as our hash
table implementation. Anyway, fewer work is always better.

Chunk regions are also split into major and minor regions with the
following rules:
  - we try to load chunks inside major region
  - we try to unload chunks outside minor region
This ensures that we will not constantly unload/load the same chunk as
the player moves back and forth between few chunks.